cleaned up shaders

+++ static/shaders/sprite.frag
@@ -1,6 +1,6 @@
 precision mediump float;
 
-varying highp vec2 v_texcoords;
+varying mediump vec2 v_texcoords;
 
 uniform sampler2D u_sprite;
 uniform vec4 u_color;
